sa_dagster/requirements.txt
agate==1.7.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==14.0
croniter==2.0.1
cryptography==42.0.3
dagster==1.6.14
dagster-cloud==1.6.14
dagster-cloud-cli==1.6.14
dagster-dbt==0.22.14
dagster-duckdb==0.22.14
dagster-graphql==1.6.14
dagster-pipes==1.6.14
dagster-webserver==1.6.14
dbt-core==1.7.10
dbt-duckdb==1.7.3
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.4.4
docstring-parser==0.15
duckdb==0.9.2
fsspec==2024.2.0
github3.py==4.0.1
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.60.1
grpcio-health-checking==1.60.1
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
leather==0.4.0
Logbook==1.5.3
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mashumaro==3.12
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
orjson==3.10.0
packaging==23.2
pandas==2.2.0
parsedatetime==2.6
pathspec==0.11.2
pendulum==3.0.0
pex==2.2.1
pluggy==1.4.0
prompt-toolkit==3.0.36
protobuf==4.25.3
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.8.0
pyodbc==5.1.0
pytest==8.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.18.0
-e git+https://github.com/inazr/starke_analytics.git@1e49275525ed6b057305774df6798cfa2e3cba6d#egg=sa_dagster&subdirectory=sa_dagster
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlglot==23.6.3
sqlglotrs==0.2.0
sqlparse==0.4.4
starlette==0.37.1
structlog==24.1.0
tabulate==0.9.0
text-unidecode==1.3
time-machine==2.13.0
tomli==2.0.1
toposort==1.10
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
universal_pathlib==0.2.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.27.1
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zipp==3.18.1
